init commit

+++ ddp_cifar100_val.py
@@ -95,7 +95,7 @@ def run_training(rank, size):
         val = torch.zeros(size=val_tensor_list[0].size())
         dist.scatter(val, scatter_list=val_tensor_list)
     else:
-        val = torch.zeros(size=(10000 / dist.get_world_size(), 4, 32, 32))
+        val = torch.zeros(size=(int(10000 / dist.get_world_size()), 4, 32, 32))
         dist.scatter(val, scatter_list=None)
     val_X = val[:, :4, :, :]
     val_y = val[:, 4, 0, 0]
